fix: Only cast int8 and unit8 when JSON marshaling arrays

+++ go/arrow/array/numeric.gen.go
@@ -101,7 +101,9 @@ func (a *Int64) MarshalJSON() ([]byte, error) {
 	vals := make([]interface{}, a.Len())
 	for i := 0; i < a.Len(); i++ {
 		if a.IsValid(i) {
-			vals[i] = float64(a.values[i]) // prevent uint8 from being seen as binary data
+
+			vals[i] = a.values[i]
+
 		} else {
 			vals[i] = nil
 		}
@@ -196,7 +198,9 @@ func (a *Uint64) MarshalJSON() ([]byte, error) {
 	vals := make([]interface{}, a.Len())
 	for i := 0; i < a.Len(); i++ {
 		if a.IsValid(i) {
-			vals[i] = float64(a.values[i]) // prevent uint8 from being seen as binary data
+
+			vals[i] = a.values[i]
+
 		} else {
 			vals[i] = nil
 		}
@@ -291,7 +295,9 @@ func (a *Float64) MarshalJSON() ([]byte, error) {
 	vals := make([]interface{}, a.Len())
 	for i := 0; i < a.Len(); i++ {
 		if a.IsValid(i) {
-			vals[i] = float64(a.values[i]) // prevent uint8 from being seen as binary data
+
+			vals[i] = a.values[i]
+
 		} else {
 			vals[i] = nil
 		}
@@ -386,7 +392,9 @@ func (a *Int32) MarshalJSON() ([]byte, error) {
 	vals := make([]interface{}, a.Len())
 	for i := 0; i < a.Len(); i++ {
 		if a.IsValid(i) {
-			vals[i] = float64(a.values[i]) // prevent uint8 from being seen as binary data
+
+			vals[i] = a.values[i]
+
 		} else {
 			vals[i] = nil
 		}
@@ -481,7 +489,9 @@ func (a *Uint32) MarshalJSON() ([]byte, error) {
 	vals := make([]interface{}, a.Len())
 	for i := 0; i < a.Len(); i++ {
 		if a.IsValid(i) {
-			vals[i] = float64(a.values[i]) // prevent uint8 from being seen as binary data
+
+			vals[i] = a.values[i]
+
 		} else {
 			vals[i] = nil
 		}
@@ -576,7 +586,9 @@ func (a *Float32) MarshalJSON() ([]byte, error) {
 	vals := make([]interface{}, a.Len())
 	for i := 0; i < a.Len(); i++ {
 		if a.IsValid(i) {
-			vals[i] = float64(a.values[i]) // prevent uint8 from being seen as binary data
+
+			vals[i] = a.values[i]
+
 		} else {
 			vals[i] = nil
 		}
@@ -671,7 +683,9 @@ func (a *Int16) MarshalJSON() ([]byte, error) {
 	vals := make([]interface{}, a.Len())
 	for i := 0; i < a.Len(); i++ {
 		if a.IsValid(i) {
-			vals[i] = float64(a.values[i]) // prevent uint8 from being seen as binary data
+
+			vals[i] = a.values[i]
+
 		} else {
 			vals[i] = nil
 		}
@@ -766,7 +780,9 @@ func (a *Uint16) MarshalJSON() ([]byte, error) {
 	vals := make([]interface{}, a.Len())
 	for i := 0; i < a.Len(); i++ {
 		if a.IsValid(i) {
-			vals[i] = float64(a.values[i]) // prevent uint8 from being seen as binary data
+
+			vals[i] = a.values[i]
+
 		} else {
 			vals[i] = nil
 		}
@@ -861,7 +877,9 @@ func (a *Int8) MarshalJSON() ([]byte, error) {
 	vals := make([]interface{}, a.Len())
 	for i := 0; i < a.Len(); i++ {
 		if a.IsValid(i) {
+
 			vals[i] = float64(a.values[i]) // prevent uint8 from being seen as binary data
+
 		} else {
 			vals[i] = nil
 		}
@@ -956,7 +974,9 @@ func (a *Uint8) MarshalJSON() ([]byte, error) {
 	vals := make([]interface{}, a.Len())
 	for i := 0; i < a.Len(); i++ {
 		if a.IsValid(i) {
+
 			vals[i] = float64(a.values[i]) // prevent uint8 from being seen as binary data
+
 		} else {
 			vals[i] = nil
 		}
